panel: separate GUI of panel settings from plugin
