+++ CMakeLists.txt
@@ -99,13 +99,13 @@ if(CLOCK_PLUGIN)
     add_subdirectory(plugin-clock)
 endif()
 
-setByDefault(COLORPICKER_PLUGIN Yes)
+setByDefault(COLORPICKER_PLUGIN No)
 if(COLORPICKER_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Color Picker")
     add_subdirectory(plugin-colorpicker)
 endif()
 
-setByDefault(CPULOAD_PLUGIN Yes)
+setByDefault(CPULOAD_PLUGIN No)
 if(CPULOAD_PLUGIN)
     find_library(STATGRAB_LIB statgrab)
 
@@ -119,7 +119,7 @@ if(CPULOAD_PLUGIN)
     endif()
 endif()
 
-setByDefault(DIRECTORYMENU_PLUGIN Yes)
+setByDefault(DIRECTORYMENU_PLUGIN No)
 if(DIRECTORYMENU_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Directory menu")
     add_subdirectory(plugin-directorymenu)
@@ -139,7 +139,7 @@ if(DESKTOPSWITCH_PLUGIN)
     add_subdirectory(plugin-desktopswitch)
 endif()
 
-setByDefault(KBINDICATOR_PLUGIN Yes)
+setByDefault(KBINDICATOR_PLUGIN No)
 if(KBINDICATOR_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Keyboard Indicator")
     add_subdirectory(plugin-kbindicator)
@@ -153,7 +153,7 @@ if(MAINMENU_PLUGIN)
     add_subdirectory(plugin-mainmenu)
 endif()
 
-setByDefault(MOUNT_PLUGIN Yes)
+setByDefault(MOUNT_PLUGIN No)
 if(MOUNT_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Mount")
     add_subdirectory(plugin-mount)
@@ -167,13 +167,13 @@ if(QUICKLAUNCH_PLUGIN)
     add_subdirectory(plugin-quicklaunch)
 endif()
 
-setByDefault(SCREENSAVER_PLUGIN Yes)
+setByDefault(SCREENSAVER_PLUGIN No)
 if(SCREENSAVER_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Screensaver")
     add_subdirectory(plugin-screensaver)
 endif()
 
-setByDefault(SENSORS_PLUGIN Yes)
+setByDefault(SENSORS_PLUGIN No)
 if(SENSORS_PLUGIN)
     find_library(SENSORS_LIB sensors)
 
@@ -195,7 +195,7 @@ if(SHOWDESKTOP_PLUGIN)
     add_subdirectory(plugin-showdesktop)
 endif()
 
-setByDefault(NETWORKMONITOR_PLUGIN Yes)
+setByDefault(NETWORKMONITOR_PLUGIN No)
 if(NETWORKMONITOR_PLUGIN)
     find_library(STATGRAB_LIB statgrab)
 
@@ -209,7 +209,7 @@ if(NETWORKMONITOR_PLUGIN)
     endif()
 endif()
 
-setByDefault(SYSSTAT_PLUGIN Yes)
+setByDefault(SYSSTAT_PLUGIN No)
 if(SYSSTAT_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "System Stats")
     add_subdirectory(plugin-sysstat)
@@ -268,7 +268,7 @@ if(VOLUME_PLUGIN)
     endif()
 endif()
 
-setByDefault(WORLDCLOCK_PLUGIN Yes)
+setByDefault(WORLDCLOCK_PLUGIN No)
 if(WORLDCLOCK_PLUGIN)
     list(APPEND STATIC_PLUGINS "worldclock")
     add_definitions(-DWITH_WORLDCLOCK_PLUGIN)
@@ -276,7 +276,7 @@ if(WORLDCLOCK_PLUGIN)
     add_subdirectory(plugin-worldclock)
 endif(WORLDCLOCK_PLUGIN)
 
-setByDefault(SPACER_PLUGIN Yes)
+setByDefault(SPACER_PLUGIN No)
 if(SPACER_PLUGIN)
     set(ENABLED_PLUGINS ${ENABLED_PLUGINS} "Spacer")
     add_subdirectory(plugin-spacer)
